+++ DicomTypeTranslation/DicomDatasetExtensions.cs
@@ -1,8 +1,10 @@
-﻿using Dicom;
+﻿
 using System;
 using System.Collections.Generic;
 using System.Data;
-using System.Text;
+
+using Dicom;
+
 
 namespace DicomTypeTranslation
 {
@@ -21,14 +23,14 @@ namespace DicomTypeTranslation
         public static DataRow ToRow(this DicomDataset dataset, DataTable inTable)
         {
             var row = inTable.Rows.Add();
-                        
+
             //for each item in the dataset
-            foreach(DicomItem i in dataset)
-                AddColumnValue(dataset,row,i);
+            foreach (DicomItem i in dataset)
+                AddColumnValue(dataset, row, i);
 
             return row;
         }
-        
+
         /// <summary>
         ///  Converts the <paramref name="dataset"/> into a new row <paramref name="inTable"/>.  Columns will automatically be created
         ///  for all top level tags that are in <paramref name="onlyTheseTags"/>.
@@ -42,9 +44,9 @@ namespace DicomTypeTranslation
             var row = inTable.Rows.Add();
 
             //for each item in the dataset
-            foreach(DicomItem i in dataset)
-                if(onlyTheseTags.Contains(DicomTypeTranslaterReader.GetColumnNameForTag(i.Tag,false)))
-                    AddColumnValue(dataset,row,i);
+            foreach (DicomItem i in dataset)
+                if (onlyTheseTags.Contains(DicomTypeTranslaterReader.GetColumnNameForTag(i.Tag, false)))
+                    AddColumnValue(dataset, row, i);
 
             return row;
         }
@@ -57,36 +59,33 @@ namespace DicomTypeTranslation
         /// <param name="inTable"></param>
         /// <param name="filterTags">Function returning true if the item should be added to the table</param>
         /// <returns></returns>
-        public static DataRow ToRow(this DicomDataset dataset, DataTable inTable, Func<DicomItem,bool> filterTags)
+        public static DataRow ToRow(this DicomDataset dataset, DataTable inTable, Func<DicomItem, bool> filterTags)
         {
             var row = inTable.Rows.Add();
 
             //for each item in the dataset
-            foreach(DicomItem i in dataset)
-                if(filterTags(i))
-                    AddColumnValue(dataset,row,i);
+            foreach (DicomItem i in dataset)
+                if (filterTags(i))
+                    AddColumnValue(dataset, row, i);
 
             return row;
 
         }
-        private static void AddColumnValue(DicomDataset dataset,DataRow row, DicomItem i)
+        private static void AddColumnValue(DicomDataset dataset, DataRow row, DicomItem i)
         {
             //get the column name for the tag
-            var name = DicomTypeTranslaterReader.GetColumnNameForTag(i.Tag,false);
+            var name = DicomTypeTranslaterReader.GetColumnNameForTag(i.Tag, false);
 
             //if we don't have it in our table yet
-            if(!row.Table.Columns.Contains(name))
+            if (!row.Table.Columns.Contains(name))
             {
                 //add it with the correct hard Type
-                var type = DicomTypeTranslater.GetNaturalTypeForVr(i.ValueRepresentation,i.Tag.DictionaryEntry.ValueMultiplicity);
-                row.Table.Columns.Add(name,type.CSharpType); 
+                var type = DicomTypeTranslater.GetNaturalTypeForVr(i.ValueRepresentation, i.Tag.DictionaryEntry.ValueMultiplicity);
+                row.Table.Columns.Add(name, type.CSharpType);
             }
 
             //populate row value
-            row[name] = DicomTypeTranslaterReader.GetCSharpValue(dataset,i);
+            row[name] = DicomTypeTranslaterReader.GetCSharpValue(dataset, i);
         }
-
-
-
     }
 }